[Kernels][benchmarks] Add target for kernel microbenchamarks and run in post-submit (#36554)
The kernel microbenchmarks were not being run in CI and they were rotting. This adds
a target to run them (check-kernel-benchmarks) and runs it in post-submit. All
tests are run with the `-t` (test) flag which runs them for a single iteration
to check for correctness only.

This runs the benchmarks using the lit driver, because it was the easiest thing
to set up right now. In the future, they should run with a dedicated benchmark
driver.

Fixes [Internal link]

KERNEL_BENCHMARKS_ORIG_REV_ID: a368eccf255ef74a06caaec5bb65b27044c2a2a7