[mojo-lang] Massively restructure parameter inference, ~NFC. (#37390)
This restructures parameter inference of function calls to be
failable and more closely follow the pattern of overload resolution.

It is deeply unfortunate that we have these parallel paths.  This
is bad for redundancy and code complexity, and also bad for compile
time.  We need to revisit all of this, but this at least pulls it
inline to be close to each other.  This is prep work for fixing

MOJO_MAX_ORIG_REV_ID: 0b14478e8f1c94ce4b2868eaba3f6aed1dddbc37