[Kernels] modernize `neon_intrinsics.mojo` (#37848)
Reworked the NEON dotprod/matmul intrinsics to eliminate a bunch of
extra wrappers and boilerplate code.

The original reason to touch this file was to extend
`_neon_dotprod_lane` to support broadcasting from a `b` that is not the
full SIMD width. The kernel I'm working on is loading `SIMD[int8, 8]`
and I want to broadcast out of that without doing some additional code
to pad the loaded vector.

LINALG_ORIG_REV_ID: c0bf7dab03dac8db3ac0e9a9459424344e0dbbfa