[mojo] Add Spin lock and Scoped lock implementation. (#38165)
This PR adds a simple spin lock and a scoped lock.
The lock implementation spins on an atomic value which represents the
owner (and -1 to represent unowned/unlocked). A parameter allows for
configuring the timeout duration.
The scoped lock enables using the spin lock in a programmer-friendly
manner by using the address of the scoped-lock object as the owner
identifier of the lock.
Usage pattern:

```
var lock = BlockingSpinLock[0.000001]()

fn inc() capturing -> Int:
     var addr = Pointer[BlockingSpinLock[0.000001]].address_of(lock)
     with BlockingScopedLock(addr) as l:
          #exclusive access to resource
```

---------

Signed-off-by: Rashid Kaleem <rashid@modular.com>
RUNTIME_ORIG_REV_ID: 0df470098af90751d6d30588857b99b211424b76