[graph-api] Pass global runtime from Mojo to MAXG (#38658)
Pass the global runtime from Mojo to the `max.graph` C API.
The issue with `getCurrentAllocator()` today is that it is coupled with
the Mojo stdlib Tensor type.
In `max.graph` we want an API to create a `dense_resource` backed by the
stdlib tensor’s memory.
And `dense_resource` needs to be able to take ownership of and free that
memory during graph compilation.
So that’s why this `tensorAttr()`  function in the MAX graph C API calls
into `getCurrentAllocator()`, which needs to return whatever allocator
is used in `KGEN_CompilerRT_AlignedAlloc()` on the Mojo side.

However, `mojo build` static links `LLCLRuntimeGlobals` so that when
the MAXG C API calls `getCurrentAllocator()` from libmof.so, it gets the
wrong global runtime.
To work around that, this PR passes the global runtime from the Mojo
side into the `max.graph` C API.

Fixes #36154.

MOJO_MAX_ORIG_REV_ID: 2b9869d51ba0f33d999915c495faf8e440c9ec61