[Kernels][mojo-stdlib] Implment failable `compile_code` (#38667)
This PR updates `compile.mojo` to use the faillable mode of
`compile_assembly`. It adds an assert to paths that access `.asm` that
asserts it succeeded. This can be used to conditionally try to compile a
function.

Fixes #37367

COMPILE_ORIG_REV_ID: 287ed8ae195a9df2859171a47ae1edc57631667d