[Kernels] add transpose_k support to CPU flash attention (#39038)
Add support to the CPU flash attention kernel to specify that the format
of the K tensor is BHSD instead of BHDS. In typical model usage, the
attention kernel is fused into a kernel that has a view op that is doing
a transpose. All of the SIMD load from the attention kernel end up being
strided accesses. This also causes problems when integrating with a KV
cache enabled kernel as the SIMD accesses end up crossing between the K
cache and K.

This first PR is not fully optimized: the packing of K is a naive
implementation. This unblocks the work to update the KV cache enabled
kernel + `llama_quantized` model to use optimal shapes for the buffers.
The next PR will fix the K packing.

I'll open an issue later to track getting the graph compiler to also set
transpose_k.

This should be largely NFC as the transpose_k=False path is largely
unchanged.

KERNEL_BENCHMARKS_ORIG_REV_ID: acdec671910ef78c1a564bcd1ce8559dd9ecf309