[Kernels][GPU] Implement powfi function that does not go through llvm (#39110)
The llvm intrinsic would not work on the GPU since we do not compile
compilerrt for it. The function is simple to implement and we already
have the implementation, so just use that and avoid the special case
and use the russian peasant algorithm in both cases.

---------

Co-authored-by: Hengjie Wang <hengjiewang@modular.com>
GPU_TEST_ORIG_REV_ID: 5600caab5cb05411af0a3e68dfb6e8680508e6d9