[KGEN] Have `compile_assembly` return an error sentinel instead of a variant (NFC) (#39687)
VariantType is being reworked, and in this case, it's possible to stuff
the error message into the string field and pass a sentinel value in the
`num_captures` field. This simplifies the API between the compiler and
the stdlib and gets rid of one load-bearing use of variant.

COMPILE_ORIG_REV_ID: 6cacef941d5359e28aa428ec12fec5e5f1dd1959