[graph-api] Add Mojo interfaces for avg_pool and max_pool ops to the Graph API (#39101)
Average and max pooling are common operations in convolutional networks,
so it may be helpful to have exposed Mojo interfaces for the
`mo.avg_pool` and `mo.max_pool` ops in the Graph API. This adds those
interfaces and some basic tests for them.

While I was at it, I removed a few stray `var`s from the previous
convolution interfaces as we are moving to `def` in the Graph API and
started to consolidate the code for calculating tensor dimensions after
filters and padding are applied.

Requires #38882 to build.

MOJO_MAX_ORIG_REV_ID: 7e745718fa59b9a9dde7f70470dbf175f4eec3c2