[Kernels] Restructuring Kernels/mojo/gpu/host/* to use a unified DLL cached context. (#39756)
Added a unique `CudaHandle` "base type" for all CUDA related handles to
avoid expensive dynamic dll calls.

This PR creates optional dependencies in the current CUDA API to a
`CudaInstance` and to the `Context` (which carries the instance). The
new API is backwards compatible with the old one, everything should
still work in the absence of a context, and DLL calls will still be
performed dynamically.

The new API can be seen in action in the `test_binary_apply.mojo` code
example.

This pattern is the foundation subsequent additions to the Context (such
as a compiled PTX cache), and allows the `Context` object to be passed
around explicitly instead of being referred through a global variable.

GPU_TEST_ORIG_REV_ID: 97e567566ebfda7d0556604bb0ea46fecbcb3275