[mojo] Introduce `@parameter for` and replace all `@unroll`
This PR deprecates and removes the `@unroll` feature and replaces with
it `@parameter for`. The `@unroll` decorator was an experiment in
compiler directives to implement performance semantics, but in practice
proved brittle. The new feature leverages the parameter system to
actually guarantee that the loop is unrolled. For real this time.

It does so by requiring that the iterator value is a parameter value,
and it also makes the induction variable a parameter value as well.

KERNEL_BENCHMARKS_ORIG_REV_ID: 0b1427fd52bc531f94c8a47a75e4f7ab08ac2540