[mojo-lang] Fix subtle issues with inferred parameter handling (#40238)
There was a hack where we changed the behaviour of parmeter binding for
a parameter based on whether a parameter inference hook was non-null.
This worked fine for implicit parameters, but inferred parameters can
come at the beginning of the parameter list. This caused issues when
binding functions outside of a call emission context, because a
parameter inference hook was not present, causing the binding behaviour
to be incorrect. This PR fixes that by carefully handling inferred and
implicit parameters and removing that hack.

This PR also happens to fix 2 extra goodies: it caught a case where an
implicit parameter *had* to be explicitly specified on an indirect
(PValue) callable. This fixes that but also enables partial binding of
indirect callables, which was something that was disabled for some
reason.

KERNEL_BENCHMARKS_ORIG_REV_ID: eb305d82b995deb6b4270dcc2daaf9630c833d26