[stdlib] Move `pow` from `math` to builtins
This brings us in line with Python where this is also a builtin.
Notably, `IntLiteral` and `FloatLiteral` does not implement this yet,
and tests will need improvement, but this change opens up the
possibility of the community contributing to these.

MATH_ORIG_REV_ID: ddde9c362eb779281207230b278dae177a7ca489