[Kernels] Use `List.resize` without value (#40694)
Recently, `List` gained an overload of `resize`
to not specify a value.  Apply this in a few places
in the kernels to improve performance.  This avoids
initializing the elements in the list with said value.
In these cases, we know the indices in the list are
about to be written to entirely, so there's no need
to fill with an initialized value only to immediately
overwrite it.

NN_ORIG_REV_ID: 482340a0bf0a430a1310d21e60d04b050af82b9d