[mojo-lang] Strengthen invariant in call emission (#40664)
The branch that adds non-trivial register-passable values to
`valuesToKeepAlive` is dead because non-trivial variadics and packs are
always emitted in-memory.

Remove the branch and replace it with an error. Also tag a new linear
issue in the relevant places.

The error can still trigger when binding a non-trivial type to
`AnyRegType`.

This removes putting `LayoutTensor` inside `StaticTuple`.

LINALG_ORIG_REV_ID: 79e0368f145683b962ffe1cbd0e9a3a66499fae2