[mojo-stdlib] Add a lifetime parameter to TaskGroup (#40877)
This lifetime parameter is supposed to be the union of the lifetimes of
all coroutines passed to `create_task`. However, we don't actually have
a way to enforce that right now (we need where statements and a
subsetting rule for lifetimes).

At the very least, add the parameter to force users to think about
lifetimes for now.

ALGORITHM_ORIG_REV_ID: 9c3fe7074809952809af631c3f56b915e355b5ba