[MAX Driver] Add data_unsafe() method to DeviceMemory (#40916)
Adds method to extract the underlying pointer from the DeviceMemory
object. This will (temporarily) be used in the KV cache example to
create an NDBuffer out of the DeviceMemory until we have a chance to
switch over to using TensorSlice, which mananages the the lifetime using
References.

Co-authored-by: Ian Tramble <ian@modular.com>
MOJO_MAX_ORIG_REV_ID: 99aa8b9c1637c17a18641368a87466c6236fcdf5