[stdlib] Move `bench_memem.mojo` to `stdlib` benchmarks (#40960)
As part of the strategy of open-sourcing the stdlib
microbenchmarks so contributors can add more them,
move the recently created `bench_memem.mojo` to a new
directory: `mojo/stdlib/benchmarks`.  Add Copybara
rules so we don't *yet* sync this benchmarks directory.

This is just a step in that direction.  This also creates a new
CMake target: `check-mojo-stdlib-benchmarks` that is part of the `check-all`
CMake target so that we continue to make sure this benchmark compiles.

In a follow-up PR, we'll rework the `lit.cfg.py` and write some scripts
so users can run these benchmarks.  Then, we can sync this directory
for Copybara rules.

Fixes MSTDL-589

KERNEL_BENCHMARKS_ORIG_REV_ID: d55faffa4142cc6d975108cbdac3b6d4ffa23c20