[mojo-stdlib] Make `TaskGroup` only take `NoneType` tasks (#41087)
In practice, TaskGroup is only used for void/NoneType tasks with no
results. This is because it's used to run a collection of tasks where
the results of the individual tasks doesn't matter much.

This simplifies the API and means in a subsequent PR, when we enable
memory-only types in async functions, we don't have to worry about
managing the memory of the results of every single task added to the
TaskGroup. This makes it more efficient overall.

ALGORITHM_ORIG_REV_ID: 5bde6018e72e6150badd5a1d61664d59f2d95b94